Refactored in js
The original script used several WSS options, instead of fetching from a CLIO server. 

It also required all trait names to be hardcoded in. This version creates the database, then updates it dynamically depending on the `trait_type` names found in the `attributes` array, making this version much more usable for any given project.

Finally, instead of requiring the user to update the `Issuer` and `Taxon` variables in the script, both are passed as arguments along with the command `node fetcher.js [Issuer] [Taxon]`.